Move shared schema types to common/schema_types
