+++ x-pack/platform/plugins/shared/lens/public/shared_components/schema_flyout/evaluate_condition.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import type { FieldDescriptor, FieldVisibilityCondition } from './types';
+import type { FieldDescriptor, FieldVisibilityCondition } from '../../../common/schema_types';
 
 interface ConditionContext {
   /** Series types from all data layers */
