+++ x-pack/platform/plugins/shared/lens/public/shared_components/schema_flyout/fields/schema_form_field.test.tsx
@@ -5,18 +5,11 @@
  * 2.0.
  */
 
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0"; you may not use this file except in compliance with the "Elastic License
- * 2.0".
- */
-
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { useForm } from 'react-hook-form';
 import { SchemaFormField } from './schema_form_field';
-import type { FieldDescriptor } from '../types';
+import type { FieldDescriptor } from '../../../../common/schema_types';
 
 const Wrapper = ({
   children,
