+++ x-pack/platform/plugins/shared/lens/public/shared_components/schema_flyout/fields/schema_form_field.tsx
@@ -5,16 +5,8 @@
  * 2.0.
  */
 
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0"; you may not use this file except in compliance with the "Elastic License
- * 2.0".
- */
-
 import React from 'react';
-import type { Control } from 'react-hook-form';
-import type { FieldDescriptor } from '../types';
+import type { SchemaFieldProps } from '../types';
 import { ToggleField } from './toggle_field';
 import { SelectField } from './select_field';
 import { NumberField } from './number_field';
@@ -24,11 +16,6 @@ import { RangeField } from './range_field';
 import { PaginationToggleField } from './pagination_toggle_field';
 import { RowHeightField } from './row_height_field';
 
-interface SchemaFormFieldProps {
-  descriptor: FieldDescriptor;
-  control: Control;
-}
-
 // Map custom widget names to base field types they should render as.
 // Custom widgets that don't have dedicated renderers yet fall back to
 // the closest generic renderer based on the schema type.
@@ -36,7 +23,7 @@ const widgetToBaseType: Record<string, string> = {
   buttonGroup: 'select',
 };
 
-export const SchemaFormField = ({ descriptor, control }: SchemaFormFieldProps) => {
+export const SchemaFormField = ({ descriptor, control }: SchemaFieldProps) => {
   // Resolve: explicit widget → mapped base type → schema type
   const widget = descriptor.widget ?? descriptor.type;
   const resolved = widgetToBaseType[widget] ?? widget;
