+++ x-pack/platform/plugins/shared/lens/public/shared_components/schema_flyout/types.ts
@@ -5,37 +5,13 @@
  * 2.0.
  */
 
-/**
- * Ready-to-render field descriptor received from the server.
- * Mirrors server/ui_schemas/types.ts FieldDescriptor.
- */
-export interface FieldDescriptor {
-  path: string;
-  type: string;
-  label: string;
-  description?: string;
-  defaultValue?: unknown;
-  options?: Array<{ value: string; label: string }>;
-  min?: number;
-  max?: number;
-  required?: boolean;
-  widget?: string;
-  props?: Record<string, unknown>;
-  tooltip?: string;
-  children?: FieldDescriptor[];
-  /** Condition that must be met for this field to be visible */
-  condition?: FieldVisibilityCondition;
-}
+import type { Control } from 'react-hook-form';
+import type { FieldDescriptor } from '../../../common/schema_types';
 
 /**
- * Conditions controlling when a field is visible.
- * All specified conditions must be met (AND logic).
+ * Shared props for all schema field components.
  */
-export interface FieldVisibilityCondition {
-  /** Show this field only when at least one layer matches one of these series types */
-  seriesTypes?: string[];
-  /** Show only when the X axis is time-based */
-  requiresTimeAxis?: boolean;
-  /** Show only when NO text-based (ES|QL) datasource is present */
-  excludeTextBased?: boolean;
+export interface SchemaFieldProps {
+  descriptor: FieldDescriptor;
+  control: Control;
 }
