+++ x-pack/platform/plugins/shared/lens/public/shared_components/schema_flyout/use_schema_descriptions.ts
@@ -7,7 +7,7 @@
 
 import { useQuery } from '@kbn/react-query';
 import { useKibana } from '@kbn/kibana-react-plugin/public';
-import type { FieldDescriptor } from './types';
+import type { FieldDescriptor } from '../../../common/schema_types';
 
 interface VizSchemaResponse {
   fields: FieldDescriptor[];
