+++ x-pack/platform/plugins/shared/lens/server/ui_schemas/types.ts
@@ -5,6 +5,8 @@
  * 2.0.
  */
 
+import type { FieldVisibilityCondition } from '../../common/schema_types';
+
 /**
  * UI schema entry — per-field overrides provided by visualization authors.
  * Acts as an allowlist: only fields listed here appear in the response.
@@ -26,41 +28,6 @@ export interface UISchemaEntry {
   options?: Array<{ value: string; label: string }>;
   /** Default value when the field is not present in the visualization state */
   defaultValue?: unknown;
-  /** Condition that must be met for this field to be visible */
-  condition?: FieldVisibilityCondition;
-}
-
-/**
- * Conditions controlling when a UI schema field is visible.
- * All specified conditions must be met (AND logic).
- */
-export interface FieldVisibilityCondition {
-  /** Show this field only when at least one layer matches one of these series types */
-  seriesTypes?: string[];
-  /** Show only when the X axis is time-based */
-  requiresTimeAxis?: boolean;
-  /** Show only when NO text-based (ES|QL) datasource is present */
-  excludeTextBased?: boolean;
-}
-
-/**
- * Ready-to-render field descriptor returned by the server endpoint.
- * The client uses this directly — no schema processing needed.
- */
-export interface FieldDescriptor {
-  path: string;
-  type: string;
-  label: string;
-  description?: string;
-  defaultValue?: unknown;
-  options?: Array<{ value: string; label: string }>;
-  min?: number;
-  max?: number;
-  required?: boolean;
-  widget?: string;
-  props?: Record<string, unknown>;
-  tooltip?: string;
-  children?: FieldDescriptor[];
   /** Condition that must be met for this field to be visible */
   condition?: FieldVisibilityCondition;
 }
